+++ packages/gov-ui-kit/src/components/progress/progress.api.ts
@@ -1,11 +1,18 @@
 import { type HTMLAttributes } from 'react';
+import { type ResponsiveAttribute } from '../../types';
+
+export type ProgressSize = 'sm' | 'md';
 
 export interface IProgressProps extends HTMLAttributes<HTMLDivElement> {
     /**
      * Size of progress component.
      * @default md
      */
-    size?: 'md' | 'sm';
+    size?: ProgressSize;
+    /**
+     * Size of the progress depending on the current breakpoint.
+     */
+    responsiveSize?: ResponsiveAttribute<ProgressSize>;
     /**
      * Current progress to be rendered.
      */