+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,9 +7,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Implement `DataList` component
+-   Handling of responsive sizes to `Progress` component
+
+### Changed
+
+-   Update `eslint` rules to enforce no circular dependencies
+
 ### Fixed
 
+-   `CardEmptyState` component to horizontally center content
+-   Spacings of `EmptyState` component when being stacked and having an object illustration
 -   Typos on documentation and comments
+
+### Removed
+
+-   `ActionItem` component
 
 ## [1.0.16] - 2024-02-27